QMP: Relax __com.redhat_drive_add parameter checking
RH-Author: Markus Armbruster <armbru@redhat.com>
Message-id: <1394040778-1544-2-git-send-email-armbru@redhat.com>
Patchwork-id: 58021
O-Subject: [PATCH 7.0 qemu-kvm 1/1] QMP: Relax __com.redhat_drive_add parameter checking
Bugzilla: 1057471
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>
RH-Acked-by: Fam Zheng <famz@redhat.com>
RH-Acked-by: Amos Kong <akong@redhat.com>
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>

I/O throttling and many new features are unavailable with this
command, because its parameter checking is overly restrictive.
Relax it.

The command was forward-ported from RHEL-6 (commit 75ad257).  It
provides access to drive_init() via QMP with the parameters restricted
to a subset of the ones recognized by drive_init().  We did that
because some parameters make sense only when configuring a frontend in
addition to a backend, and the command doesn't do that.

The parameter filtering is implemented as a whitelist.  The
forward-port neglected to update the whitelist for all the stuff that
has changed since RHEL-6.

Due to new features like driver-specific parameters, a whitelist is no
longer convenient.  Replace by a blacklist that contains exactly the
drive_init() parameters that are already filtered out on RHEL-6.

Signed-off-by: Markus Armbruster <armbru@redhat.com>
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>