fix vga type for older machines
RH-Author: Dr. David Alan Gilbert <dgilbert@redhat.com>
Message-id: <1466070273-5859-4-git-send-email-dgilbert@redhat.com>
Patchwork-id: 70634
O-Subject: [RHEL-7.3 qemu-kvm-rhev PATCH v2 3/3] fix vga type for older machines
Bugzilla: 1344320
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Eduardo Habkost <ehabkost@redhat.com>
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Paolo Bonzini <pbonzini@redhat.com>

From: "Dr. David Alan Gilbert" <dgilbert@redhat.com>

7.1.0 and older use cirrus.

In the 7.2.0 and older worlds this came from the .default_display
field that was only set in rhel720 (to std) and the others
used the default.  But the new qemu default is std where
we want our older types to use cirrus so we can't leave it blank;
be explicit.

Signed-off-by: Dr. David Alan Gilbert <dgilbert@redhat.com>
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>