STX: mig-compat: set proper compat flags for pc-i440fx-2.2 mach type
The upstream version of qemu-kvm-ev enabled "modern
mode" for virtio-pci.

We need to disable the use of "modern mode" for the pc-i440fx-2.2
machine type in order to make it behave like earlier releases,
otherwise live migration fails.  This is similar to what RHEL does
for the pc-i440fx-rhel7.2.0 machine type.

Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>